terraform/python/estimate/requirements.txt
agraffe
fastapi
google-cloud-logging
numpy
pandas-datareader
polars
pyarrow
scipy
